fix scroll jump issue
